requirements.txt
numpy
pandas
matplotlib
scipy
opencv-python-headless
redis
scikit-learn
streamlit
streamlit-webrtc
insightface
onnxruntime
jupyter
python-dotenv